--- test/contract/auditing/test.yml
@@ -1,6 +1,0 @@
-apiVersion: atlas.mongodb.com/v1
-kind: AtlasProject
-metadata:
-  name: my-project
-spec:
-  name: Test Atlas Operator Project
