+++ test/int/integration_suite_test.go
@@ -76,6 +76,7 @@ var (
 )
 
 func TestAPIs(t *testing.T) {
+	control.SkipTestUnless(t, "AKO_INT_TEST")
 
 	utilruntime.Must(scheme.AddToScheme(scheme.Scheme))
 	utilruntime.Must(akov2.AddToScheme(scheme.Scheme))
